+++ solr/__init__.py
@@ -7,3 +7,5 @@ from .parsers import parse_json_response, parse_query_response
 from .utils import read_response, check_response_status
 from .schema import SchemaAPI
 from .mlt import MoreLikeThis
+from .suggest import Suggest
+from .response import SpellcheckResult